* DBReply 핸들러 추가 ** Helper는 의미상 아닌거 같아서 Handler로 변경 * packetId를 빼지 않고 sessionId 먼저 빼낸 것 수정 * packetId를 WORD로 사용했던 것 수정

+++ RIO_Test/PacketHandler.cpp
@@ -32,7 +32,7 @@ bool PacketManager::HandlePacket(RIOTestSession& session, EchoStringPacket& pack
 bool PacketManager::HandlePacket(RIOTestSession& session, CallTestProcedurePacket& packet)
 {
 	CSerializationBuf& buffer = *CSerializationBuf::Alloc();
-	WORD packetId = DBServerProtocol::PACKET_ID::TEST;
+	UINT packetId = DBServerProtocol::PACKET_ID::TEST;
 	buffer << packetId << session.GetSessionId() << packet.id3;
 	buffer.WriteBuffer((char*)packet.testString, sizeof(packet.testString));
 
@@ -44,7 +44,7 @@ bool PacketManager::HandlePacket(RIOTestSession& session, CallTestProcedurePacke
 bool PacketManager::HandlePacket(RIOTestSession& session, CallSelectTest2ProcedurePacket& packet)
 {
 	CSerializationBuf& buffer = *CSerializationBuf::Alloc();
-	WORD packetId = DBServerProtocol::PACKET_ID::SELECT_TEST_2;
+	UINT packetId = DBServerProtocol::PACKET_ID::SELECT_TEST_2;
 	buffer << packetId << session.GetSessionId() << packet.id;
 
 	DBClient::GetInstance().CallProcedure(buffer);
